Split off `API` and `Server` for recycle-ics
