+++ recycle-ics/src/Recycle/Ics.hs
@@ -1,5 +1,4 @@
 {-# LANGUAGE DataKinds #-}
-{-# LANGUAGE TypeOperators #-}
 
 module Recycle.Ics
   ( main,
@@ -14,42 +13,25 @@ import qualified Data.ByteString.Lazy.Char8 as BSL
 import qualified Data.ByteString.Lazy.Char8 as BSL8
 import Data.Foldable (for_)
 import Data.IORef (newIORef)
-import Data.Text (Text)
 import qualified Data.Text as T
-import Data.Time hiding (getZonedTime)
 import Network.HTTP.Client.TLS
   ( newTlsManagerWith,
     tlsManagerSettings,
   )
-import Network.HTTP.Media.MediaType
-import Network.Wai.Application.Static
-  ( defaultWebAppSettings,
-  )
 import Network.Wai.Handler.Warp
 import Network.Wai.Middleware.RequestLogger
-import Numeric.Natural (Natural)
 import Paths_recycle_ics
 import Recycle.AppM
 import Recycle.Class
 import Recycle.Ics.ICalendar
 import Recycle.Ics.Opts
-import Recycle.Types
-import Servant.API
-import Servant.API.QueryParamForm
+import Recycle.Ics.Server
 import Servant.Client
   ( BaseUrl (..),
     Scheme (..),
     mkClientEnv,
   )
-import Servant.Server
-  ( Handler,
-    ServerT,
-    hoistServer,
-    serve,
-  )
-import Servant.Server.StaticFiles (serveDirectoryWith)
 import System.FilePath
-import WaiAppStatic.Types
 
 main :: IO ()
 main = do
@@ -93,109 +75,4 @@ main = do
       putStrLn "Starting server"
       run port
         . logStdoutDev
-        . serve pRecycleIcsAPI
-        . hoistServer pRecycleIcsAPI (recycleToHandler env)
-        $ recycleIcsServer wwwDir
-
-recycleToHandler :: Env -> RecycleM a -> Handler a
-recycleToHandler env act = liftIO $ act `runRecycle` env
-
-calculateDateRange :: HasTime m => DateRange -> m (Range Day)
-calculateDateRange = \case
-  AbsoluteDateRange r -> pure r
-  RelativeDateRange relRange -> do
-    today <- localDay . zonedTimeToLocalTime <$> getZonedTime
-    pure
-      Range
-        { rangeFrom = addDays (rangeFrom relRange) today,
-          rangeTo = addDays (rangeTo relRange) today
-        }
-
--- brittany-disable-next-binding
-type RecycleIcsAPI =
-  ( "api"
-      :> ( "search-zipcode"
-             :> QueryParam' '[Required] "q" (SearchQuery Natural)
-             :> UVerb 'GET '[JSON] '[WithStatus 200 [FullZipcode]]
-             :<|> "search-street"
-               :> QueryParam' '[Required] "zipcode" ZipcodeId
-               :> QueryParam' '[Required] "q" (SearchQuery Text)
-               :> UVerb 'GET '[JSON] '[WithStatus 200 [Street]]
-             :<|> "generate"
-               :> QueryParamForm DateRange
-               :> QueryParam' '[Required] "lc" LangCode
-               :> QueryParamForm FractionEncoding
-               :> QueryParam' '[Required] "z" ZipcodeId
-               :> QueryParam' '[Required] "s" StreetId
-               :> QueryParam' '[Required] "hn" HouseNumber
-               :> UVerb 'GET '[ICalendar] '[WithStatus 200 BSL.ByteString]
-         )
-      :<|> Raw
-  )
-
-data ICalendar
-
-instance Accept ICalendar where
-  contentType Proxy = "text" // "calendar"
-
-instance MimeRender ICalendar BSL.ByteString where
-  mimeRender Proxy = id
-
-pRecycleIcsAPI :: Proxy RecycleIcsAPI
-pRecycleIcsAPI = Proxy
-
-recycleIcsServer ::
-  forall m.
-  (HasRecycleClient m, HasTime m) =>
-  FilePath ->
-  ServerT RecycleIcsAPI m
-recycleIcsServer dataDir =
-  (searchZipcode :<|> searchStreet :<|> generateCollection) :<|> serveWww
-  where
-    searchZipcode ::
-      SearchQuery Natural -> m (Union '[WithStatus 200 [FullZipcode]])
-    searchZipcode query = do
-      zipcodes <- searchZipcodes (Just query)
-      pure . Z . I $ WithStatus @200 zipcodes
-    searchStreet ::
-      ZipcodeId -> SearchQuery Text -> m (Union '[WithStatus 200 [Street]])
-    searchStreet zipcodeId query = do
-      streets <- searchStreets (Just zipcodeId) (Just query)
-      pure . Z . I $ WithStatus @200 streets
-    generateCollection ::
-      DateRange ->
-      LangCode ->
-      FractionEncoding ->
-      ZipcodeId ->
-      StreetId ->
-      HouseNumber ->
-      m (Union '[WithStatus 200 BSL8.ByteString])
-    generateCollection collectionQueryDateRange collectionQueryLangCode collectionQueryFractionEncoding collectionQueryZipcode collectionQueryStreet collectionQueryHouseNumber =
-      do
-        let collectionQuery = CollectionQuery {..}
-        collections <- runCollectionQuery collectionQuery
-        pure . Z . I . WithStatus @200 $ printVCalendar collections
-    serveWww =
-      serveDirectoryWith
-        (defaultWebAppSettings dataDir)
-          { ssRedirectToIndex = True,
-            ssIndices = [unsafeToPiece "index.html"]
-          }
-
-runCollectionQuery ::
-  (HasRecycleClient m, HasTime m) => CollectionQuery -> m VCalendar
-runCollectionQuery CollectionQuery {..} = do
-  range <- calculateDateRange collectionQueryDateRange
-  collections <-
-    getCollections
-      collectionQueryZipcode
-      collectionQueryStreet
-      collectionQueryHouseNumber
-      range
-  -- for_ collections $ liftIO . print
-
-  pure $
-    mkVCalendar
-      collectionQueryLangCode
-      collectionQueryFractionEncoding
-      collections
+        $ recycleIcsApp wwwDir env
